pyproject.toml
[project]
name = "transcripty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.10"
dependencies = [
    "flet>=0.24.1",
    "openai-whisper>=20240930",
    "pvrecorder>=1.2.3",
]

[project.scripts]
main = "transcripty.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
